+++ scripts/Data_SpecialEventData.rb
@@ -1,4 +1,16 @@
 class SpecialEventData
+  attr_reader :flags
+  attr_reader :collision
+
+  def initialize(flags, collision)
+    @flags = flags
+    @collision = collision
+  end
+
+  def self.get(name)
+    return SPECIAL_EVENTS[name]
+  end
+
   SPECIAL_EVENTS = {
     :bigpool => SpecialEventData.new(
       [:bottom],
