--- scripts/Name.rb
@@ -1,3 +1,0 @@
-def enter_name
-  $game_temp.name_calling = true
-end
