+++ scripts/Puzzle_Pixel.rb
