+++ scripts/Sprite_Light.rb
